pyproject.toml
[tool.poetry]
name = "mondaysdk"
version = "0.1.0"
description = ""
authors = ["Jonathan Crum <jcrum@theobogroup.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
strawberry-graphql = {extras = ["debug-server"], version = "^0.144.2"}
rich = "^12.6.0"
fastapi = "^0.88.0"
pendulum = "^2.1.2"
beartype = "^0.11.0"
pydantic = "^1.10.2"
pyjwt = "^2.6.0"
python-dotenv = "^1.0.0"
asgiref = "^3.6.0"
pytest = "^7.3.1"
devtools = "^0.11.0"
httpx = "^0.24.0"
trio = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
